Divide plotting module into 6 distinct modules
This should make it much easier to find code for specific types
of plots and will be a first step toward consolidating plotting
code that is redundant between plot routines.
