+++ mpas_analysis/__main__.py
@@ -48,7 +48,7 @@ from mpas_analysis.shared import AnalysisTask
 from mpas_analysis.shared.analysis_task import \
     update_time_bounds_from_file_names
 
-from mpas_analysis.shared.plot.plotting import _register_custom_colormaps, \
+from mpas_analysis.shared.plot.colormap import _register_custom_colormaps, \
     _plot_color_gradients
 
 from mpas_analysis import ocean
