+++ mpas_analysis/ocean/plot_depth_integrated_time_series_subtask.py
@@ -19,7 +19,7 @@ import xarray
 
 from mpas_analysis.shared import AnalysisTask
 
-from mpas_analysis.shared.plot.plotting import timeseries_analysis_plot
+from mpas_analysis.shared.plot import timeseries_analysis_plot
 
 from mpas_analysis.shared.io import open_mpas_dataset, write_netcdf
