+++ mpas_analysis/ocean/time_series_ocean_regions.py
@@ -20,7 +20,7 @@ from multiprocessing.pool import ThreadPool
 
 from mpas_analysis.shared.analysis_task import AnalysisTask
 
-from mpas_analysis.shared.plot.plotting import timeseries_analysis_plot
+from mpas_analysis.shared.plot import timeseries_analysis_plot
 
 from mpas_analysis.shared.io import open_mpas_dataset, write_netcdf
