+++ mpas_analysis/shared/plot/__init__.py
@@ -1,0 +1,10 @@
+from mpas_analysis.shared.plot.time_series import timeseries_analysis_plot, \
+    timeseries_analysis_plot_polar
+
+from mpas_analysis.shared.plot.climatology_map import plot_polar_comparison, \
+    plot_global_comparison, plot_polar_projection_comparison
+
+from mpas_analysis.shared.plot.vertical_section import \
+    plot_vertical_section_comparison, plot_vertical_section
+
+from mpas_analysis.shared.plot.oned import plot_1D
